pyproject.toml
[project]
name = "reltio-open-mcp-server"
version = "0.1.0"
description = "Reltio Open Source MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.10.1",
    "requests>=2.32.3",
    "urllib3>=2.4.0",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.10.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing"
markers = [
    "asyncio: mark a test as an asyncio test",
]
